requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic_settings
python-multipart
passlib[bcrypt]
pydantic[email]
pytest
asyncpg
PyJWT[crypto]
aiosqlite
pytest-asyncio